python-ai/requirements.txt
flask
opencv-python
mysql-connector-python
requests
python-dotenv